Fix trait filtering (GitHub vs. Bitbucket)
Currently there is the problem, that the GitHub behaviour is also
selectable for Bitbucket sources and vice versa.

After this change the traits for GitHub are only visible for GitHub and
the traits for Bitbucket are only visible for Bitbucket sources.

The overridden Descriptor methods are the same as the ones in the traits in
the plugins "cloudbees-bitbucket-branch-source" and "github-branch-source".